Allow desired signifance to be set
`significance` query param controls whether a highlighted winner is shown.  If the winner's signifiance is greater than the significance param, it gets highlighted.

Significance defaults to 90% for peramlinks without a setting, and for new queries.

Currently the input form is hidden by default, and becomes visible when you click the animated chart.

Closes #8